Q&A: Clarify both client and AP must be patched
Some sources were claiming only the AP *or* the client needs to be
patched. For example:
https://exchange.xforce.ibmcloud.com/collection/396ecb6880625d6e58dd7636b7c8e8fd
They claim: ".. if even only one of the devices (client or access
point) has been patched, the pair are not vulnerable to this form
of attack." However both the client (4-way handshake) and the AP
(fast BSS transition) must be patched.